refactor(cardinal): move store manager into package
